Code cleanup.
Moslty removal of unused code/classes.

+++ dev/core/src/com/google/gwt/dev/util/msg/FormatterForInteger.java
@@ -22,6 +22,6 @@ public final class FormatterForInteger extends Formatter {
 
   @Override
   public String format(Object toFormat) {
-    return ((Integer) toFormat).toString();
+    return toFormat.toString();
   }
 }
